requirements.txt
pdfminer.six==20240706
pypdf==3.17.0